+++ Examiner/Models/Answeers.cs
@@ -12,6 +12,6 @@
 
         public string SelectedAnsweer { get; set; } = "";
 
-        public byte Mark { get; set; }
+        public int Mark { get; set; }
     }
 }
